openrr_apps

Struct TeleopPluginConfig

pub struct TeleopPluginConfig {
    pub path: PathBuf,
    pub args: Option<String>,
    pub args_from_path: Option<PathBuf>,
}

Fields
path: PathBuf

Path to the plugin. If no extension is specified, the default extension for cdylib on the current OS will be selected. (linux: .so, macos: .dylib, windows: .dll)

args: Option<String>

Arguments passed when creating this instance.

args_from_path: Option<PathBuf>

Pass the contents of the specified file as an argument.